# app/Http/Controllers/Api/User/UserController.php
<?php

namespace App\Http\Controllers\Api\User;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Hash;
use Illuminate\Validation\Rule;
use App\Models\User;

class UserController extends Controller
{
    /**
     * Récupérer le profil de l'utilisateur connecté
     */

        /**
     * @OA\Get(
     *     path="/api/user/profile",
     *     summary="Profil utilisateur",
     *     description="Récupère les informations du profil de l'utilisateur connecté",
     *     tags={"User"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Profil récupéré avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="user", type="object",
     *                 @OA\Property(property="id", type="integer", example=1),
     *                 @OA\Property(property="name", type="string", example="John Doe"),
     *                 @OA\Property(property="user_name", type="string", example="johndoe"),
     *                 @OA\Property(property="email", type="string", example="john@example.com"),
     *                 @OA\Property(property="indicatif", type="string", example="+225"),
     *                 @OA\Property(property="contact", type="string", example="01234567"),
     *                 @OA\Property(property="commune", type="string", example="Abobo"),
     *                 @OA\Property(property="CMU", type="string", nullable=true, example="CMU12345"),
     *                 @OA\Property(property="profile_picture", type="string", nullable=true),
     *                 @OA\Property(property="actif", type="integer", example=1),
     *                 @OA\Property(property="created_at", type="string", format="date-time"),
     *                 @OA\Property(property="updated_at", type="string", format="date-time")
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Non authentifié"
     *     )
     * )
     */

    public function profile(Request $request)
    {

        $user = $request->user();
        return response()->json([
            'status' => 'success',
            'user' => $user
        ]);
    }

    /**
     * Mettre à jour le profil
     */

        /**
     * @OA\Put(
     *     path="/api/user/profile",
     *     summary="Mettre à jour le profil",
     *     description="Met à jour les informations du profil utilisateur",
     *     tags={"User"},
     *     security={{"bearerAuth":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="name", type="string", maxLength=191, example="John Doe Updated"),
     *             @OA\Property(property="user_name", type="string", maxLength=191, example="johndoe_updated"),
     *             @OA\Property(property="email", type="string", format="email", maxLength=191, example="john.updated@example.com"),
     *             @OA\Property(property="indicatif", type="string", maxLength=10, example="+225"),
     *             @OA\Property(property="contact", type="string", maxLength=191, example="07654321"),
     *             @OA\Property(property="commune", type="string", maxLength=191, example="Cocody"),
     *             @OA\Property(property="CMU", type="string", maxLength=191, nullable=true, example="CMU67890"),
     *             @OA\Property(property="profile_picture", type="string", maxLength=191, nullable=true)
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Profil mis à jour avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="user", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Erreur de validation"
     *     )
     * )
     */

    public function updateProfile(Request $request)
    {
        $user = $request->user();

        $validatedData = $request->validate([
        
        ]);

        $user->fill($validatedData);
        $user->save();

        $user->refresh();

        return response()->json([
            'status' => 'success',
            'user' => $user
        ]);
    }

    /**
     * Changer le mot de passe
     */


        /**
     * @OA\Post(
     *     path="/api/user/change-password",
     *     summary="Changer le mot de passe",
     *     description="Change le mot de passe de l'utilisateur connecté",
     *     tags={"User"},
     *     security={{"bearerAuth":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"current_password","password","password_confirmation"},
     *             @OA\Property(property="current_password", type="string", format="password", example="oldpassword123"),
     *             @OA\Property(property="password", type="string", format="password", minLength=6, example="newpassword123"),
     *             @OA\Property(property="password_confirmation", type="string", format="password", example="newpassword123")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Mot de passe changé avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="message", type="string", example="Mot de passe changé")
     *         )
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Mot de passe actuel incorrect",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="message", type="string", example="Le mot de passe actuel est incorrect.")
     *         )
     *     )
     * )
     */

    public function changePassword(Request $request)
    {
        $request->validate([
            'current_password' => 'required|string',
            'password' => 'required|string|min:6|confirmed',
        ]);

        $user = $request->user();

        if (!Hash::check($request->current_password, $user->password)) {
            return response()->json([
                'status' => 'error',
                'message' => 'Le mot de passe actuel est incorrect.'
            ], 400);
        }

        $user->update([
            'password' => Hash::make($request->password)
        ]);

        return response()->json([
            'status' => 'success',
            'message' => 'Mot de passe changé'
        ]);
    }

    /**
     * Mettre à jour les préférences de notifications
     */


        /**
     * @OA\Put(
     *     path="/api/user/notifications",
     *     summary="Préférences de notifications",
     *     description="Met à jour les préférences de notifications de l'utilisateur",
     *     tags={"User"},
     *     security={{"bearerAuth":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"notifications"},
     *             @OA\Property(property="notifications", type="boolean", example=true)
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Préférences mises à jour",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="message", type="string", example="Préférences de notifications mises à jour"),
     *             @OA\Property(property="notifications", type="boolean", example=true)
     *         )
     *     )
     * )
     */
    public function updateNotifications(Request $request)
    {
        $request->validate([
            'notifications' => 'required|boolean'
        ]);

        $user = $request->user();
        $user->update([
            'notifications' => $request->notifications
        ]);

        return response()->json([
            'status' => 'success',
            'message' => 'Préférences de notifications mises à jour',
            'notifications' => $user->notifications
        ]);
    }
}