# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E_Messe API",
        "description": "API pour l'application E_Messe - Gestion des messes",
        "contact": {
            "email": "leprodev03@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur Local"
        },
        {
            "url": "https://votre-domaine.com/api",
            "description": "Serveur de Production"
        }
    ],
    "paths": {
...
        "/api/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Profil utilisateur",
                "description": "Récupère les informations du profil de l'utilisateur connecté",
                "operationId": "643f3fba69c3b90788166059a9db5094",
                "responses": {
                    "200": {
                        "description": "Profil récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "user_name": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "indicatif": {
                                                    "type": "string",
                                                    "example": "+225"
                                                },
                                                "contact": {
                                                    "type": "string",
                                                    "example": "01234567"
                                                },
                                                "commune": {
                                                    "type": "string",
                                                    "example": "Abobo"
                                                },
                                                "CMU": {
                                                    "type": "string",
                                                    "example": "CMU12345",
                                                    "nullable": true
                                                },
                                                "profile_picture": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "actif": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Mettre à jour le profil",
                "description": "Met à jour les informations du profil utilisateur",
                "operationId": "788178f3bf26d10d78a92a30e51734ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "example": "John Doe Updated"
                                    },
                                    "user_name": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "example": "johndoe_updated"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 191,
                                        "example": "john.updated@example.com"
                                    },
                                    "indicatif": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "+225"
                                    },
                                    "contact": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "example": "07654321"
                                    },
                                    "commune": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "example": "Cocody"
                                    },
                                    "CMU": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "example": "CMU67890",
                                        "nullable": true
                                    },
                                    "profile_picture": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Changer le mot de passe",
                "description": "Change le mot de passe de l'utilisateur connecté",
                "operationId": "51ca9b37aa2f939aff779e61958697bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "oldpassword123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6,
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe changé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mot de passe changé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Mot de passe actuel incorrect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Le mot de passe actuel est incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/notifications": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Préférences de notifications",
                "description": "Met à jour les préférences de notifications de l'utilisateur",
                "operationId": "f95d1285bb419bae7867db50fd5ab2bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "notifications"
                                ],
                                "properties": {
                                    "notifications": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Préférences mises à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Préférences de notifications mises à jour"
                                        },
                                        "notifications": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}